% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_test.R
\name{crosses_centromere}
\alias{crosses_centromere}
\title{Does it Cross the Centromere?}
\usage{
crosses_centromere(chr, start, end, centromeres)
}
\arguments{
\item{chr}{The chromosome that the region belongs to}

\item{start}{The start position of the region}

\item{end}{The end position of the region}

\item{centromeres}{Centromere positions from get_centromere_regions()}
}
\description{
This function checks whether a region crosses the centromere
}
