% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_signatures.R
\name{get_reference_signatures}
\alias{get_reference_signatures}
\title{Retrieving Reference Mutation Signatures}
\usage{
get_reference_signatures(path = NULL)
}
\arguments{
\item{path}{String indicating path to reference signatures table file. Defaults to bundled 30-signatures.}
}
\description{
This function imports reference mutation signatures to serve as a comparison point when
performing non-negative least squares decomposition.
}
