% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_test.R
\name{get_subtelomere_coordinate}
\alias{get_subtelomere_coordinate}
\title{Get Subtelomere Coordinate}
\usage{
get_subtelomere_coordinate(chr, subtelomere, startOrEnd)
}
\arguments{
\item{chr}{The chromosome of interest}

\item{subtelomere}{A subtelomeres object from get_subtelomere_regions()}

\item{startOrEnd}{Takes on value either "start" or "end" depending on which position to return}
}
\description{
Convenience function that retrieves the start or end coordinate of a specific subtelomere
}
