% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ranges.R
\name{import_ranges}
\alias{import_ranges}
\title{Imports ranges output from APOLLOH}
\usage{
import_ranges(path, col.names, sanitization_function = sanitize_granges)
}
\arguments{
\item{path}{Path to LOH ranges file}

\item{col.names}{A vector providing the column names of the LOH ranges file}

\item{sanitization_function}{A callback function which takes a GRanges object as input and returns a sanitized version}
}
\description{
This function returns a GRanges object based on the data contained in a TSV.
This TSV can be output from APOLLOH or similar program.
}
