% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh_test.R
\name{loh_test}
\alias{loh_test}
\title{The HRD-LOH Test}
\usage{
loh_test(gr)
}
\arguments{
\item{gr}{GRanges object obtained from import_ranges()}
}
\description{
This function runs the Telomeric Allelic Imbalance test (HRD-LOH).
}
