% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_catalog.R
\name{mutation_catalog}
\alias{mutation_catalog}
\title{Obtain the Mutational Catalog / Spectrum}
\usage{
mutation_catalog(vr)
}
\arguments{
\item{vr}{VRanges file, obtained from import_snvs()}
}
\description{
This function returns the mutational catalogue / mutational spectrum of a cancer genome.
}
