% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnls_exposures.R
\name{nnls_exposures_worker}
\alias{nnls_exposures_worker}
\title{NNLS Worker Function}
\usage{
nnls_exposures_worker(subjectMotifs, refSignature, fractions)
}
\arguments{
\item{subjectMotifs}{Mutation Catalog from get_mutation_catalog()}

\item{refSignature}{Reference signatures from get_reference_signatures()}

\item{fractions}{If TRUE, will return exposures as a fraction of total mutation burden}
}
\description{
Performs the NNLS calculation itself.
}
