% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_snv.R
\name{run_snv}
\alias{run_snv}
\title{Deciphers SNV Signatures}
\usage{
run_snv(vcf_file = NULL, snv_file = NULL, genome = "hg19",
  fractions = FALSE, silent = FALSE)
}
\arguments{
\item{vcf_file}{Path to an VCF file, suitable for import_vcfs()}

\item{snv_file}{Path to an SNV file, suitable for import_snvs()}

\item{genome}{ID of the genome being used (default: 'hg19')}

\item{fractions}{If TRUE, mutation signature exposures are normalized to sum to 1 (default: 'FALSE')}

\item{silent}{If TRUE, does not print exposures output (default: 'FALSE')}
}
\description{
This function loads an SNV data file and runs the mutation signature deciphering
process on it. It returns the output of nnls_exposures.
}
