% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multiple.R
\name{test_multiple}
\alias{test_multiple}
\title{Run the HRD Tests on Multiple Files}
\usage{
test_multiple(loh_files, output_file, genome = "hg19",
  number_of_cores = NULL, previous_output = NULL, log_path = NULL)
}
\arguments{
\item{output_file}{Path to an output file where a table of the results will be stored.}

\item{genome}{ID of the genome being used (default: 'hg19')}

\item{number_of_cores}{Specifies number of cores to run across. If NULL, then determines number of cores automatically. (default: NULL)}

\item{previous_output}{Path to output file previously generated by test_multiple. If provided, will only append new cases and not update existing ones. (default: NULL)}

\item{snv_file}{Character vector of paths to SNV files}
}
\description{
This is a convenience function to run_test() on multiple files.
}
